% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_structure_operators.R
\name{changeGrade}
\alias{changeGrade}
\title{Grade operator: add or subtract}
\usage{
changeGrade(grade, delta)
}
\arguments{
\item{grade}{a string containing the current grade in the form \code{G?}, where \code{?} is a number.}

\item{delta}{a number containing the relative change in grade to apply. \code{0} retains the current grade as-is.}
}
\value{
a string containing the new grade.
}
\description{
\code{\link{changeGrade}} is an operator for grade values.
}
\examples{
changeGrade("G4",  0) ## G4
changeGrade("G4",  1) ## G5
changeGrade("G4", -1) ## G3
changeGrade("G10", 1) ## G11

}
