% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_structure_operators.R
\name{changePhase}
\alias{changePhase}
\title{Phase operator: move to next phase}
\usage{
changePhase(phase, assessment_structure)
}
\arguments{
\item{phase}{a string containing the current phase in the format \code{P?}, where \code{?} is a number.}

\item{assessment_structure}{an \code{\linkS4class{assessment_structure}} object.}
}
\value{
a string containing the new phase.
}
\description{
\code{\link{changePhase}} is an operator for phase values.
}
\examples{
## assessment uses two phases
changePhase("P1", assessment_structure_math) ## P2
changePhase("P2", assessment_structure_math) ## P1

}
