% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_structure_operators.R
\name{changeTest}
\alias{changeTest}
\title{Test operator: move to next phase}
\usage{
changeTest(test, phase, assessment_structure)
}
\arguments{
\item{test}{a string containing the current test in the format \code{T?}, where \code{?} is a number.}

\item{phase}{a string containing the current phase in the format \code{P?}, where \code{?} is a number.}

\item{assessment_structure}{an \code{\linkS4class{assessment_structure}} object.}
}
\value{
a string containing the new test.
}
\description{
\code{\link{changeTest}} is an operator for test values.
}
\examples{
## assessment uses two phases
changeTest("T1", "P1", assessment_structure_math) ## T1
changeTest("T1", "P2", assessment_structure_math) ## T2

}
