% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.config.MAAT.R
\name{convert.samples.MAAT}
\alias{convert.samples.MAAT}
\title{Convert samples for MAAT}
\usage{
convert.samples.MAAT(trait.samples)
}
\arguments{
\item{trait.samples}{a matrix or dataframe of samples from the trait distribution}
}
\value{
matrix or dataframe with values transformed
}
\description{
convert parameters and parameter names from PEcAn database default units/names with MAAT
}
\details{
Performs model specific unit conversions on a a list of trait values,
such as those provided to write.config
}
\author{
Shawn Serbin, Anthony Walker
}

