% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{createModule}
\alias{createModule}
\title{Create a single module}
\usage{
createModule(constraints, item_pool, item_attrib, passage_attrib)
}
\arguments{
\item{constraints}{constraints data. A \code{\link{data.frame}} or a csv file name to be used in \code{\link[TestDesign]{loadConstraints}}.}

\item{item_pool}{item pool data. A \code{\link{data.frame}} or a csv file name to be used in \code{\link[TestDesign]{loadItemPool}}.}

\item{item_attrib}{item attribute data. A \code{\link{data.frame}} or a csv file name to be used in \code{\link[TestDesign]{loadItemAttrib}}.}

\item{passage_attrib}{passage attribute data. A \code{\link{data.frame}} or a csv file name to be used in \code{\link[TestDesign]{loadStAttrib}}.}
}
\value{
a \code{\linkS4class{module}} object.
}
\description{
\code{\link{createModule}} is a function for creating a \code{\linkS4class{module}} object
based on the item pool, attribute, and constraints.
}
