% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examinee_class.R
\docType{class}
\name{examinee_list-class}
\alias{examinee_list-class}
\title{Class 'examinee_list': a list of examinees}
\description{
\code{\linkS4class{examinee_list}} is an S4 class to represent a list of examinees.
}
\section{Slots}{

\describe{
\item{\code{examinee_list}}{a list of \code{\linkS4class{examinee}} objects.}

\item{\code{assessment_structure}}{an \code{\linkS4class{assessment_structure}} object.}

\item{\code{is_complete}}{\code{TRUE} after being updated from running \code{\link{maat}}. (default = \code{FALSE})}
}}

