% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examinee_class.R
\docType{methods}
\name{examinee_list-operators}
\alias{examinee_list-operators}
\alias{[,examinee_list,numeric-method}
\alias{[,examinee_list,numeric,ANY,ANY-method}
\title{Basic operators for examinee list objects}
\usage{
\S4method{[}{examinee_list,numeric}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{an \code{\linkS4class{examinee_list}} object.}

\item{i}{examinee indices to use in subsetting.}

\item{j, drop, ...}{not used, exists for compatibility.}
}
\description{
Create a subset of an \code{\linkS4class{examinee_list}} object:

\itemize{
  \item{\code{examinee_list[i]}}
}
}
\examples{
examinee_list_math[1:2]

}
