% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getAdaptivityIndex}
\alias{getAdaptivityIndex}
\title{Calculate adaptivity indices from an examinee list object}
\usage{
getAdaptivityIndex(examinee_list)
}
\arguments{
\item{examinee_list}{a list containing \code{\linkS4class{examinee}} objects, returned from \code{\link{thyCAT}}.}
}
\value{
a data frame containing adaptivity indices by test and also for all tests combined.
}
\description{
\code{\link{getAdaptivityIndex}} is a function for calculating adaptivity indices from the output of \code{\link{thyCAT}}.
}
