% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getAdministeredItemsPerTest}
\alias{getAdministeredItemsPerTest}
\title{Get administered items per test}
\usage{
getAdministeredItemsPerTest(examinee_list)
}
\arguments{
\item{examinee_list}{a list containing \code{\linkS4class{examinee}} objects, returned from \code{\link{thyCAT}}.}
}
\value{
a list containing administered items in each test and also for all tests combined.
}
\description{
\code{\link{getAdministeredItemsPerTest}} is a function for extracting the administered items stored in the
\code{\linkS4class{examinee}} objects.
}
