% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getItemExposureRate}
\alias{getItemExposureRate}
\title{Get item exposure rates from an examinee list}
\usage{
getItemExposureRate(examinee_list, module_list)
}
\arguments{
\item{examinee_list}{an \code{\linkS4class{examinee_list}} object from \code{\link{simExaminees}}, returned from \code{\link{maat}}.}

\item{module_list}{a module list from \code{\link{loadModules}}.}
}
\value{
the table of item exposure rate.
}
\description{
\code{\link{getItemExposureRate}} is a function for building an item exposure rate table.
}
