% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getItemNamesPerGrade}
\alias{getItemNamesPerGrade}
\title{Get item names per grade}
\usage{
getItemNamesPerGrade(module_list)
}
\arguments{
\item{module_list}{a module list from \code{\link{loadModules}}.}
}
\value{
item names per grade.
}
\description{
\code{\link{getItemNamesPerGrade}} is a function for extracting item names from a module list.
}
\examples{
getItemNamesPerGrade(module_list_math)

}
