% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_structure_operators.R
\name{getRelativeGrade}
\alias{getRelativeGrade}
\title{Grade operator: difference between two grades}
\usage{
getRelativeGrade(current_grade, initial_grade)
}
\arguments{
\item{current_grade}{a string containing the current grade in the form \code{G?}, where \code{?} is a number.}

\item{initial_grade}{a string containing the initial grade in the form \code{G?}, where \code{?} is a number.}
}
\value{
the grade difference of the current grade relative to the initial grade.
}
\description{
\code{\link{getRelativeGrade}} is an operator for grade values.
}
\examples{
getRelativeGrade("G4", "G3") ## 1
getRelativeGrade("G5", "G3") ## 2
getRelativeGrade("G2", "G3") ## -1

}
