% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getSE}
\alias{getSE}
\title{Calculate standard error from an examinee list object}
\usage{
getSE(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_maat}} object from \code{\link{maat}}.}
}
\value{
a list containing SE by test and also for all tests combined.
}
\description{
\code{\link{getSE}} is a function for calculating the standard error of the estimates.
}
