% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_class.R
\docType{class}
\name{module-class}
\alias{module-class}
\title{Class 'module': a module}
\description{
\code{\linkS4class{module}} is an S4 class to represent a module.
}
\section{Slots}{

\describe{
\item{\code{module_id}}{the ID of the module.}

\item{\code{constraints}}{a \code{\link[TestDesign:constraints-class]{constraints}} object.}
}}

