% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{module_list_math}
\alias{module_list_math}
\alias{assessment_structure_math}
\alias{examinee_list_math}
\alias{cut_scores_math}
\title{Example item pools}
\description{
Example data for a 6-module assessment.
}
\details{
\itemize{
  \item \code{assessment_structure_math} an \code{\linkS4class{assessment_structure}} object defining 3 tests with 2 phases in each test. Also defines routing limits as \code{G - 1} and \code{G + 2}, where \code{G} is the starting grade.
  \item \code{examinee_list_math} a list of \code{\linkS4class{examinee}} objects. The number of examinees is 10. This can be created using \code{\link{simExaminees}}.
  \item \code{module_list_math} a list of \code{\linkS4class{module}} objects. This can be created using \code{\link{loadModules}}.
  \item \code{cut_scores_math} a list of theta cut scores. This is used in the \code{cut_scores} argument of the \code{\link{maat}} function.
}
}
\keyword{datasets}
