% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotModuleRoutes}
\alias{plotModuleRoutes}
\title{Plot module routes}
\usage{
plotModuleRoutes(
  examinee_list,
  assessment_structure,
  examinee_id = "all",
  font_size = 15,
  box_colour = "PaleTurquoise"
)
}
\arguments{
\item{examinee_list}{the output from \code{\link{maat}}.}

\item{assessment_structure}{an \code{\linkS4class{assessment_structure}} object.}

\item{examinee_id}{the examinee ID to plot the module route. \code{all} plots the count that each routing node was used. (default = \code{all})}

\item{font_size}{the font size for the count labels. (default = \code{15})}

\item{box_colour}{the colour for the boxes. (default = \code{PaleTurquoise})}
}
\description{
\code{\link{plotModuleRoutes}} is a function for building a chart for module routes.
}
