% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/print_functions.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,module-method}
\title{Extension of print()}
\usage{
\S4method{print}{module}(x)
}
\arguments{
\item{x}{an object to display the content.}
}
\description{
Extension of print()
}
