% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.MAAT.R
\name{process_maat_variable}
\alias{process_maat_variable}
\title{Prepare single MAAT output variable for PEcAn standard}
\usage{
process_maat_variable(data, oldname, newname, oldunits, newunits, longname,
  ncdims, missval = -999)
}
\arguments{
\item{data}{Data values, as vector, matrix, or array}

\item{oldname}{Old variable name (character)}

\item{newname}{Target variable name (character)}

\item{oldunits}{Old variable units (character). Passed into [PEcAn.utils::misc.convert].}

\item{newunits}{New variable units (character). Passed into [PEcAn.utils::misc.convert].}

\item{longname}{Long, descriptive variable name (character)}

\item{ncdims}{NetCDF dimensions (list). Passed into `dim` argument
of [ncdf4::ncvar_def]}

\item{missval}{Value for missing data (numeric, default = -999)}
}
\value{
List containing NetCDF variable (`var`) and converted data
  values (`dat`)
}
\description{
Prepare single MAAT output variable for PEcAn standard
}
\author{
Alexey Shiklomanov
}
