% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administered_functions.R
\name{removeItemData}
\alias{removeItemData}
\title{Remove item data from examinee list}
\usage{
removeItemData(examinee_list)
}
\arguments{
\item{examinee_list}{a list containing \code{\linkS4class{examinee}} objects.}
}
\value{
a list containing \code{\linkS4class{examinee}} objects,
with \code{item_data} data stripped for compact storage.
}
\description{
\code{\link{removeItemData}} is a function to remove the item data from
the \code{\linkS4class{examinee}} objects for the reduction of file size.
}
