% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/show_functions.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,module-method}
\title{Extension of show()}
\usage{
\S4method{show}{module}(object)
}
\arguments{
\item{object}{an object to display the content.}
}
\description{
Extension of show()
}
