% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examinee_updaters.R
\name{updateItemData}
\alias{updateItemData}
\title{Update the item data slot of an examinee object}
\usage{
updateItemData(examinee_object, module_position, solution)
}
\arguments{
\item{examinee_object}{an \code{\linkS4class{examinee}} object.}

\item{module_position}{the current module position.}

\item{solution}{an \code{\link[TestDesign:output_Shadow_all-class]{output_Shadow_all}} object.}
}
\value{
an \code{\linkS4class{examinee}} object with its \code{item_data} slot updated.
}
\description{
\code{\link{updateItemData}} is a function for updating \code{\linkS4class{examinee}} objects after completing a module.
}
\details{
\code{\link{updateItemData}} updates the \code{item_data} slot with an \code{\link[TestDesign:item_pool-class]{item_pool}} object
that contains administered items in the module.
}
