% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examinee_updaters.R
\name{updateLog}
\alias{updateLog}
\title{Update the routing log of an examinee object}
\usage{
updateLog(examinee_object, current_module_position)
}
\arguments{
\item{examinee_object}{an \code{\linkS4class{examinee}} object.}

\item{current_module_position}{the current module position.}
}
\value{
an \code{\linkS4class{examinee}} object with its \code{grade_log}, \code{phase_log}, \code{test_log}, and \code{module_log} slots updated.
}
\description{
\code{\link{updateLog}} is a function for updating \code{\linkS4class{examinee}} objects after completing a module.
\code{\link{updateLog}} updates logs with grades, phases, tests and modules.
}
