% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examinee_updaters.R
\name{updateModule}
\alias{updateModule}
\title{Update the current module of an examinee object}
\usage{
updateModule(examinee_object, module_list)
}
\arguments{
\item{examinee_object}{an \code{\linkS4class{examinee}} object.}

\item{module_list}{a module list from \code{\link{loadModules}}.}
}
\value{
an \code{\linkS4class{examinee}} object with its \code{current_module} slot updated.
}
\description{
\code{\link{updateModule}} is a function for updating \code{\linkS4class{examinee}} objects after completing a module.
\code{\link{updateModule}} assigns an \code{\linkS4class{module}} object from the supplied list to match the grade and the phase the \code{\linkS4class{examinee}} is in.
}
