% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getAdaptivityIndex}
\alias{getAdaptivityIndex}
\title{Calculate adaptivity indices from an examinee list object}
\usage{
getAdaptivityIndex(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_maat}} object from \code{\link{maat}}.}
}
\value{
a data frame containing adaptivity indices by test and also for all tests combined.
}
\description{
\code{\link{getAdaptivityIndex}} is a function for calculating adaptivity indices from the output of \code{\link{maat}}.
}
