% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getBias}
\alias{getBias}
\title{Calculate bias from an examinee list object}
\usage{
getBias(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_maat}} object from \code{\link{maat}}.}
}
\value{
a list containing bias by test and also for all tests combined.
}
\description{
\code{\link{getBias}} is a function for calculating the bias of ability estimates of the simulation results.
}
