% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getItemExposureRate}
\alias{getItemExposureRate}
\title{Get item exposure rates from an examinee list}
\usage{
getItemExposureRate(examinee_list, module_list)
}
\arguments{
\item{examinee_list}{an examinee list.}

\item{module_list}{a module list from \code{\link{loadModules}}.}
}
\description{
\code{\link{getItemExposureRate}} is a function for building an item exposure rate table.
}
