% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getItemExposureRate}
\alias{getItemExposureRate}
\title{Get item exposure rates from an examinee list}
\usage{
getItemExposureRate(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_maat}} object from \code{\link{maat}}.}
}
\value{
the table of item exposure rate.
}
\description{
\code{\link{getItemExposureRate}} is a function for building an item exposure rate table.
}
