% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getRMSE}
\alias{getRMSE}
\title{Calculate RMSE from an examinee list object}
\usage{
getRMSE(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_maat}} object from \code{\link{maat}}.}
}
\value{
a list containing RMSE by test and also for all tests combined.
}
\description{
\code{\link{getRMSE}} is a function for calculating root mean square error (RMSE)
for the simulation results.
}
