% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.MAAT.R
\name{met2model.MAAT}
\alias{met2model.MAAT}
\title{Create MAAT met driver files}
\usage{
met2model.MAAT(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  leap_year = TRUE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs (CF met drivers) are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where MAAT met outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{leap_year}{Enforce Leap-years? If set to TRUE, will require leap years to have 366 days. 
If set to false, will require all years to have 365 days. Default = TRUE.}

\item{...}{additional arguments, currently ignored}
}
\description{
met2model wrapper for MAAT
}
\author{
Shawn P. Serbin
}
