% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.MAAT.R
\name{model2netcdf.MAAT}
\alias{model2netcdf.MAAT}
\title{Function to convert MAAT model output to standard netCDF format}
\usage{
model2netcdf.MAAT(rundir, outdir, sitelat = -999, sitelon = -999,
  start_date = NULL, end_date = NULL)
}
\arguments{
\item{rundir}{Location of MAAT model run (i.e. MAAT project) directory with all required model run inputs.
This is needed to identify model runs with and without met drivers and control the model output conversion process}

\item{outdir}{Location of MAAT model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert MAAT output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\examples{
\dontrun{
example.output <- system.file("out.csv",package="PEcAn.MAAT")
model2netcdf.MAAT(rundir="~/scratch/run/", outdir="~/", sitelat=9.154, sitelon=-79.848, 
start_date="2014-01-01 00:00:00", end_date="2014-12-31 00:00:00")
}

}
\author{
Shawn Serbin, Anthony Walker
}
