% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model2netcdf.MAAT.R
\name{model2netcdf.MAAT}
\alias{model2netcdf.MAAT}
\title{Function to convert MAAT model output to standard netCDF format}
\usage{
model2netcdf.MAAT(outdir, sitelat = -999, sitelon = -999,
  start_date = NULL, end_date = NULL)
}
\arguments{
\item{outdir}{Location of MAAT model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert MAAT output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
Shawn Serbin, Anthony Walker
}

