% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\docType{class}
\name{output_maat-class}
\alias{output_maat-class}
\title{Class 'output_maat': a simulation output}
\description{
\code{\linkS4class{output_maat}} is an S4 class to represent a simulation output.
}
\section{Slots}{

\describe{
\item{\code{examinee_list}}{a list of \code{\linkS4class{examinee}} objects.}

\item{\code{assessment_structure}}{an \code{\linkS4class{assessment_structure}} object.}
}}

