% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\docType{class}
\name{output_maat-class}
\alias{output_maat-class}
\title{Class 'output_maat': a simulation output}
\description{
\code{\linkS4class{output_maat}} is an S4 class to represent a simulation output.
}
\section{Slots}{

\describe{
\item{\code{examinee_list}}{a list of \code{\linkS4class{examinee}} objects.}

\item{\code{assessment_structure}}{an \code{\linkS4class{assessment_structure}} object.}

\item{\code{module_list}}{a module list from \code{\link{loadModules}}.}

\item{\code{config}}{the list of \code{\linkS4class{config_Shadow}} objects used in the simulation for each module.}

\item{\code{cut_scores}}{the cut scores used in the simulation.}

\item{\code{overlap_control_policy}}{the policy used in the simulation.}

\item{\code{transition_policy}}{the policy used in the simulation.}

\item{\code{combine_policy}}{the policy used in the simulation.}

\item{\code{transition_CI_alpha}}{the transition parameter used in the simulation.}

\item{\code{transition_percentile_lower}}{the transition parameter used in the simulation.}

\item{\code{transition_percentile_upper}}{the transition parameter used in the simulation.}

\item{\code{initial_theta_list}}{the starting theta values used in the simulation.}

\item{\code{prior_mean_policy}}{the policy used in the simulation.}

\item{\code{prior_mean_user}}{the prior parameters used in the simulation.}

\item{\code{prior_sd}}{the prior parameters used in the simulation.}
}}

