% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/plot_functions.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,examinee-method}
\alias{plot,list-method}
\title{Extension of plot()}
\usage{
\S4method{plot}{examinee}(x, y, cut_scores)

\S4method{plot}{list}(x, y, cut_scores, theta_range = c(-4, 4), main = NULL)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{cut_scores}{a named list containing cut scores for each grade.}

\item{theta_range}{the theta range to use in scatter plots when \code{x} is an examinee list.}

\item{main}{the figure title to use in scatter plots when \code{x} is an examinee list.}
}
\description{
Extension of plot()
}
\examples{
library(TestDesign)
config <- createShadowTestConfig(
  final_theta = list(
    method = "MLE"
  )
)
examinee_list <- maat(
  examinee_list          = examinee_list_math,
  assessment_structure   = assessment_structure_math,
  module_list            = module_list_math,
  overlap_control_policy = "all",
  transition_CI_alpha    = 0.05,
  config                 = config,
  cut_scores             = cut_scores_math
)
examinee <- examinee_list[[1]]
plot(examinee, cut_scores = cut_scores_math)

}
