% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotModuleRoutes}
\alias{plotModuleRoutes}
\title{Plot module routes}
\usage{
plotModuleRoutes(
  examinee_list,
  examinee_id = "all",
  font_size = 15,
  box_color = "PaleTurquoise"
)
}
\arguments{
\item{examinee_list}{an \code{\linkS4class{examinee_list}} object from \code{\link{simExaminees}}, returned from \code{\link{maat}}.}

\item{examinee_id}{the examinee ID to plot the module route. \code{all} plots the count that each routing node was used. (default = \code{all})}

\item{font_size}{the font size for arrow labels. (default = \code{15})}

\item{box_color}{the fill color for boxes. (default = \code{PaleTurquoise})}
}
\description{
\code{\link{plotModuleRoutes}} is a function for building a chart for module routes.
}
