% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_functions.R
\name{updateAssessmentLevelTheta}
\alias{updateAssessmentLevelTheta}
\title{Update the assessment-level theta of an examinee object}
\usage{
updateAssessmentLevelTheta(examinee_object, config)
}
\arguments{
\item{examinee_object}{an \code{\linkS4class{examinee}} object.}

\item{config}{a \code{\linkS4class{config_Shadow}} object. The \code{final_theta} slot is used.}
}
\value{
an \code{\linkS4class{examinee}} object with its \code{estimated_theta} slot updated.
}
\description{
\code{\link{updateAssessmentLevelTheta}} is a function for updating \code{\linkS4class{examinee}} objects after completing all modules.
\code{\link{updateAssessmentLevelTheta}} computes the assessment-level theta. Estimation options are based on the \code{final_theta} slot of the \code{config} object.
}
