% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_functions.R
\name{updateThetaUsingCombined}
\alias{updateThetaUsingCombined}
\title{Update theta estimates using combined responses from a test}
\usage{
updateThetaUsingCombined(examinee_object, current_module_position, config)
}
\arguments{
\item{examinee_object}{an \code{\linkS4class{examinee}} object.}

\item{current_module_position}{the current module position.}

\item{config}{a \code{\linkS4class{config_Shadow}} object. The config for obtaining final estimates is used.}
}
\value{
an \code{\linkS4class{examinee}} object with its \code{estimated_theta_by_test} slot updated.
}
\description{
\code{\link{updateThetaUsingCombined}} is a function for updating \code{\linkS4class{examinee}} objects after completing a module.
\code{\link{updateThetaUsingCombined}} adds final theta estimates using all administered items in the test. A test may consist of multiple phases.
}
